All bot commands working. No testing yet.

+++ src/rankflix-console/src/main/java/pt/graca/menu/RankflixMenuService.java
@@ -1,69 +1,67 @@
 package pt.graca.menu;
 
-import pt.graca.domain.Media;
-import pt.graca.domain.User;
-import pt.graca.service.RankflixService;
-import pt.graca.service.exceptions.InvalidRatingException;
+import pt.graca.api.domain.Media;
+import pt.graca.api.domain.User;
+import pt.graca.api.service.RankflixService;
+import pt.graca.api.service.exceptions.review.InvalidRatingException;
 import pt.graca.discord.DiscordWebhookService;
-import pt.graca.service.external.PastebinService;
+import pt.graca.infra.generator.IRankGenerator;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Scanner;
 
-public class RankflixMenuService extends Menu {
+public class RankflixMenuService extends ConsoleMenu {
 
-    public RankflixMenuService(Scanner scanner, RankflixService service, PastebinService pastebinService, DiscordWebhookService discordWebhookService) {
+    public RankflixMenuService(Scanner scanner, RankflixService service, IRankGenerator rankGenerator) {
         super(scanner);
 
         this.service = service;
-        this.pastebinService = pastebinService;
-        this.discordWebhookService = discordWebhookService;
+        this.rankGenerator = rankGenerator;
     }
 
     private final RankflixService service;
-    private final PastebinService pastebinService;
-    private final DiscordWebhookService discordWebhookService;
+    private final IRankGenerator rankGenerator;
+    private final DiscordWebhookService discordWebhookService
+            = new DiscordWebhookService(System.getenv("RANKFLIX_DISCORD_WEBHOOK_URL"));
 
-    @MenuOption("Create a new user")
+    @ConsoleMenuOption("Create a new user")
     public void createUser() throws Exception {
         String username = read("Enter the username:");
         service.createUser(username, null);
     }
 
-    @MenuOption("Add rating")
-    public void addRating() throws Exception {
+    @ConsoleMenuOption("Add review")
+    public void addReview() throws Exception {
         String username = read("Enter the username: ");
         User user = service.findUserByUsername(username);
 
         String mediaTmdbId = read("Enter the media's TMDB id: ");
 
-        String rating = read("Enter the rating: ");
+        String rating = read("Enter the averageRating: ");
         try {
-            service.addRating(user.id, Integer.parseInt(mediaTmdbId), Float.parseFloat(rating), null);
+            service.addReview(user.id, Integer.parseInt(mediaTmdbId), Float.parseFloat(rating), null);
         } catch (NumberFormatException e) {
             throw new InvalidRatingException(rating);
         }
     }
 
-    @MenuOption("Delete rating (only for after 5 minutes)")
-    public void deleteRating() throws Exception {
+    @ConsoleMenuOption("Delete review (only for after 5 minutes)")
+    public void deleteReview() throws Exception {
         String username = read("Enter the username: ");
         User user = service.findUserByUsername(username);
+        if (user == null) throw new NoSuchElementException("User not found");
 
         int mediaTmdbId = Integer.parseInt(read("Enter the media's TMDB id: "));
-        service.findRankedMediaByTmdbId(mediaTmdbId);
 
         service.deleteRating(mediaTmdbId, user.id);
     }
 
-    @MenuOption("Send ranking to Discord webhook")
+    @ConsoleMenuOption("Send ranking to Discord webhook")
     public void sendToDiscord() throws Exception {
         List<Media> media = service.getTopRankedMedia(null);
 
-        // var chartUrl = chartService.generateRankingChart(media);
-        var pasteUrl = pastebinService.createPaste(media);
-
-        // send to discord
-        discordWebhookService.sendMessage(pasteUrl);
+        var generatedRankUrl = rankGenerator.generateRankUrl(media);
+        discordWebhookService.sendMessage(generatedRankUrl);
     }
 }
